CertificateManager object, watch channel sender in renew certificate
